BAH-4754 refactor changes
