+++ apps/command-palette/src/models/commandPaletteConfig.ts
