+++ apps/home/tsconfig.lib.json
@@ -42,10 +42,10 @@
       "path": "../../packages/bahmni-widgets/tsconfig.lib.json"
     },
     {
-      "path": "../command-palette/tsconfig.lib.json"
+      "path": "../../packages/bahmni-services/tsconfig.lib.json"
     },
     {
-      "path": "../../packages/bahmni-services/tsconfig.lib.json"
+      "path": "../command-palette/tsconfig.lib.json"
     },
     {
       "path": "../../packages/bahmni-design-system/tsconfig.lib.json"
