+++ distro/src/main.tsx
@@ -29,7 +29,6 @@ applyBahmniTheme(BAHMNI_DEFAULT_THEME);
 initFontAwesome();
 
 const root = createRoot(document.getElementById('root') as HTMLElement);
-
 root.render(
   <StrictMode>
     <BrowserRouter basename={PUBLIC_PATH ?? '/'}>
