+++ packages/bahmni-services/src/i18n/index.ts
@@ -1,8 +1,8 @@
+// Export the default initialized i18n instance
 export { default as initAppI18n } from './i18n';
 export { useTranslation } from 'react-i18next';
 export {
   getUserPreferredLocale,
-  getMergedTranslations,
   normalizeTranslationKey,
   extractObservationFormTranslations,
   type ObservationFormTranslations,
