+++ packages/bahmni-services/src/index.ts
@@ -5,7 +5,6 @@ export {
   useTranslation,
   normalizeTranslationKey,
   getUserPreferredLocale,
-  getMergedTranslations,
 } from './i18n';
 export { useCamera } from './cameraService';
 export {
