+++ packages/bahmni-widgets/src/commandPalette/CommandPalette.tsx
@@ -1,81 +1,13 @@
-import { type PatientSearchResult } from '@bahmni/services';
 import { Command } from 'cmdk';
 import React, { useCallback, useState } from 'react';
 import { createPortal } from 'react-dom';
 import { useTranslation } from 'react-i18next';
-import {
-  useCommandPalette,
-  type PatientAction,
-  type PatientActionContext,
-  type PatientFieldKey,
-  type PatientFieldsConfig,
-  type SearchAnnotation,
-} from './CommandPaletteContext';
+import type { SearchAnnotation } from './models';
+import { PatientCommandItem } from './PatientCommandItem';
 import styles from './styles/CommandPalette.module.scss';
+import { useCommandPalette } from './useCommandPalette';
 import { useCommandPaletteSearch } from './useCommandPaletteSearch';
-
-interface FieldDef {
-  labelKey: string;
-  getValue: (p: PatientSearchResult) => string | null | undefined;
-}
-
-const FIELD_LABEL_KEYS = {
-  name: 'COMMAND_PALETTE_FIELD_NAME',
-  identifier: 'COMMAND_PALETTE_FIELD_ID',
-  age: 'COMMAND_PALETTE_FIELD_AGE',
-  gender: 'COMMAND_PALETTE_FIELD_GENDER',
-  birthDate: 'COMMAND_PALETTE_FIELD_DOB',
-  addressFieldValue: 'COMMAND_PALETTE_FIELD_ADDRESS',
-  extraIdentifiers: 'COMMAND_PALETTE_FIELD_EXTRA_IDS',
-  customAttribute: 'COMMAND_PALETTE_FIELD_ATTRIBUTE',
-  activeVisitUuid: 'COMMAND_PALETTE_FIELD_ACTIVE_VISIT',
-} as const satisfies Record<PatientFieldKey, string>;
-
-const PATIENT_FIELD_MAP: Record<PatientFieldKey, FieldDef> = {
-  name: {
-    labelKey: FIELD_LABEL_KEYS.name,
-    getValue: (p) =>
-      [p.givenName, p.middleName, p.familyName].filter(Boolean).join(' '),
-  },
-  identifier: {
-    labelKey: FIELD_LABEL_KEYS.identifier,
-    getValue: (p) => p.identifier,
-  },
-  age: {
-    labelKey: FIELD_LABEL_KEYS.age,
-    getValue: (p) => p.age,
-  },
-  gender: {
-    labelKey: FIELD_LABEL_KEYS.gender,
-    getValue: (p) => p.gender,
-  },
-  birthDate: {
-    labelKey: FIELD_LABEL_KEYS.birthDate,
-    getValue: (p) => (p.birthDate ? String(p.birthDate) : null),
-  },
-  addressFieldValue: {
-    labelKey: FIELD_LABEL_KEYS.addressFieldValue,
-    getValue: (p) => p.addressFieldValue,
-  },
-  extraIdentifiers: {
-    labelKey: FIELD_LABEL_KEYS.extraIdentifiers,
-    getValue: (p) => p.extraIdentifiers,
-  },
-  customAttribute: {
-    labelKey: FIELD_LABEL_KEYS.customAttribute,
-    getValue: (p) => p.customAttribute,
-  },
-  activeVisitUuid: {
-    labelKey: FIELD_LABEL_KEYS.activeVisitUuid,
-    getValue: (p) => (p.activeVisitUuid ? 'Active' : null),
-  },
-};
-
-const filterItems = (value: string, search: string): number => {
-  if (value.startsWith('patient:')) return 1;
-  if (search.length < 2) return 1;
-  return value.toLowerCase().includes(search.toLowerCase()) ? 1 : 0;
-};
+import { filterItems } from './utils';
 
 const SearchIcon: React.FC = () => (
   <svg
@@ -94,150 +26,6 @@ const SearchIcon: React.FC = () => (
     />
   </svg>
 );
-
-function getInitials(givenName: string, familyName: string): string {
-  const first = givenName?.[0] ?? '';
-  const last = familyName?.[0] ?? '';
-  return (first + last).toUpperCase() || '?';
-}
-
-function buildPrimaryText(
-  patient: PatientSearchResult,
-  primaryFields: PatientFieldKey[],
-): string {
-  return primaryFields
-    .map((key) => PATIENT_FIELD_MAP[key].getValue(patient))
-    .filter(Boolean)
-    .join(' · ');
-}
-
-interface PatientCommandItemProps {
-  patient: PatientSearchResult;
-  patientFieldsConfig: PatientFieldsConfig;
-  patientActions: PatientAction[];
-  activeActionIndex: number;
-  onNavigate: (path: string) => void;
-}
-
-const PatientCommandItem: React.FC<PatientCommandItemProps> = ({
-  patient,
-  patientFieldsConfig,
-  patientActions,
-  activeActionIndex,
-  onNavigate,
-}) => {
-  const { t } = useTranslation('command-palette');
-  const [isExpanded, setIsExpanded] = useState(false);
-  const activeAction = patientActions[activeActionIndex];
-
-  const fullName = PATIENT_FIELD_MAP.name.getValue(patient) ?? '';
-  const initials = getInitials(patient.givenName, patient.familyName);
-  const primaryText = buildPrimaryText(
-    patient,
-    patientFieldsConfig.primaryFields,
-  );
-  const patientCtx: PatientActionContext = {
-    patientUuid: patient.uuid,
-    patientIdentifier: patient.identifier ?? '',
-  };
-
-  const additionalFields = patientFieldsConfig.additionalFields
-    .map((key) => ({
-      label: t(PATIENT_FIELD_MAP[key].labelKey),
-      value: PATIENT_FIELD_MAP[key].getValue(patient),
-    }))
-    .filter((f) => f.value);
-
-  const hasAdditional = additionalFields.length > 0;
-
-  const toggleExpanded = (e: React.SyntheticEvent) => {
-    e.stopPropagation();
-    e.preventDefault();
-    setIsExpanded((prev) => !prev);
-  };
-
-  const handleChevronKeyDown = (e: React.KeyboardEvent) => {
-    if (e.key === ' ' || e.key === 'Enter') toggleExpanded(e);
-  };
-
-  return (
-    <Command.Item
-      value={`patient:${patient.uuid}:${fullName} ${patient.identifier}`}
-      className={styles.item}
-      data-patient-uuid={patient.uuid}
-      onSelect={() => {
-        if (!activeAction) return;
-        onNavigate(activeAction.getPath(patientCtx));
-      }}
-    >
-      <span className={styles.avatar} aria-hidden="true">
-        {initials}
-      </span>
-
-      <span className={styles.itemContent}>
-        <span className={styles.itemLabel}>{primaryText}</span>
-        {isExpanded && hasAdditional && (
-          <span className={styles.additionalFields}>
-            {additionalFields.map((f) => (
-              <span key={f.label} className={styles.additionalField}>
-                <span className={styles.fieldLabel}>{f.label}:</span> {f.value}
-              </span>
-            ))}
-          </span>
-        )}
-      </span>
-
-      {hasAdditional && (
-        <button
-          className={styles.chevronButton}
-          onClick={toggleExpanded}
-          onKeyDown={handleChevronKeyDown}
-          aria-label={
-            isExpanded
-              ? t('COMMAND_PALETTE_COLLAPSE_DETAILS')
-              : t('COMMAND_PALETTE_EXPAND_DETAILS')
-          }
-          aria-expanded={isExpanded}
-          tabIndex={0}
-          type="button"
-        >
-          <span className={styles.chevronIcon}>{isExpanded ? '▲' : '▼'}</span>
-        </button>
-      )}
-
-      {patientActions.length > 0 && (
-        <span className={styles.itemActions}>
-          {patientActions.map((action, index) => {
-            const isActive = index === activeActionIndex;
-            return (
-              <button
-                key={action.id}
-                className={`${styles.actionButton} ${isActive ? styles.actionButtonDefault : ''}`}
-                onClick={(e) => {
-                  e.stopPropagation();
-                  onNavigate(action.getPath(patientCtx));
-                }}
-                title={
-                  isActive
-                    ? t('COMMAND_PALETTE_ACTION_ENTER_HINT', {
-                        label: t(action.label, { defaultValue: action.label }),
-                      })
-                    : t(action.label, { defaultValue: action.label })
-                }
-                type="button"
-              >
-                <span className={styles.actionButtonLabel}>
-                  {t(action.label, { defaultValue: action.label })}
-                </span>
-                {isActive && <kbd className={styles.actionEnterHint}>↵</kbd>}
-              </button>
-            );
-          })}
-        </span>
-      )}
-    </Command.Item>
-  );
-};
 
 export const CommandPalette: React.FC = () => {
   const { t } = useTranslation('command-palette');
