+++ packages/bahmni-widgets/src/commandPalette/index.ts
@@ -1,15 +1,13 @@
 export { CommandPaletteProvider } from './CommandPaletteProvider';
-export type { CommandPaletteProviderProps } from './CommandPaletteProvider';
-export { useCommandPalette } from './CommandPaletteContext';
+export { useCommandPalette } from './useCommandPalette';
+export type { CommandPaletteContextType } from './CommandPaletteContext';
 export type {
   AnnotationSearchType,
-  CommandPaletteContextType,
   NavItem,
   PatientAction,
   PatientActionContext,
+  PatientFieldKey,
   PatientFieldsConfig,
   SearchAnnotation,
   TriggerConfig,
-} from './CommandPaletteContext';
-export { PATIENT_FIELD_KEYS } from './commandPaletteConfigSchema';
-export type { PatientFieldKey } from './commandPaletteConfigSchema';
+} from './models';
